+++ src/app/components/Header/Header.css
@@ -1,0 +1,20 @@
+.header-container {
+    width: 100%;
+    display: flex;
+    justify-content: space-between;
+    align-items: baseline;
+    margin-top: 29px;
+}
+
+.header-container img {
+    margin-right: 10px;
+}
+
+.header-logo {
+    margin-left: 7rem;
+}
+
+.header-social-media {
+    margin-right: 6rem;
+}
+
